backend/requirements.txt
Django==4.0.4
mysqlclient==2.1.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.1.0